Add files via upload
The project, titled "Excel Interactive Dashboard," aims to leverage data visualization and analysis tools to construct an interactive dashboard using Microsoft Excel. The primary dataset, "bike_buyers Main Dataset.xlsx," forms the foundation of this project. Through the implementation of advanced Excel features and functions, the dashboard will offer stakeholders an intuitive and visually appealing platform to explore and comprehend key insights derived from the dataset. The interactive nature of the dashboard will help users to dynamically filter, sort, and navigate through the data, facilitating a deeper understanding of the underlying information. Overall, this project is the showcase of my skills in excel.